Move unit tests to new file
